Remove need for posix_ipc
Use python's own shared memory concepts. This involves some ugly workarounds, but should also be safer
and opens up the possibility of pynbody working on windows
