+++ pynbody/__init__.py
@@ -78,6 +78,6 @@ configuration.configure_snapshot_and_halo_loading_priority()
 
 derived_array = snapshot.simsnap.SimSnap.derived_quantity
 
-__version__ = '1.5.0'
+__version__ = '1.5.1'
 
 __all__ = ['load', 'new', 'derived_array']
