+++ pynbody/default_config.ini
@@ -422,3 +422,9 @@ path: /path/to/camb
 
 FoF-ignore: SF, NSF, Stars
 Sub-ignore: GrNr, FirstSubOfHalo, SubParentHalo, SubMostBoundID, InertiaTensor, SF, NSF, NsubPerHalo, Stars
+
+[shared-array]
+
+# delay in seconds between detecting a buffer is no longer needed and releasing the underlying memory
+# This delay is necessitated by implementation details (see pynbody.array.shared)
+cleanup-delay: 1e-2
