+++ setup.py
@@ -159,7 +159,6 @@ install_requires = [
     'h5py>=2.10.0',
     'matplotlib>=3.0.0',
     'numpy>=1.21.6',
-    'posix_ipc>=0.8',
     'scipy>=1.0.0'
 ]
